port tests to tape, introduce strict, initial package config
