+++ lib/keeper.js
