+++ lib/store.js
@@ -1,66 +1,16 @@
 "use strict";
 
-var _ = require("lodash");
-
-/**
- * Shared prototype of all Stores, e.g. baseclass.
- */
-var StoreStateBasePrototype = {
-    getState: function getState() {
-        return this.state;
-    },
-
-    clearState: function(){
-        for ( var key in this.state ){
-            this.state[key] = null;
-        }
-    },
-
-    observe: function(attr, handler) {
-        if (!this.__listeners[attr]) {
-            this.__listeners[attr] = [];
-        }
-
-        this.__listeners[attr].push(handler);
-    },
-
-    stopObserving: function(attr, handler) {
-        if (!this.__listeners[attr])
-            throw "Fail to unbind " + attr + " for store " + this.storeName;
-            
-        if (!this.__listeners[attr].indexOf(handler) === -1)
-            return console.warn("Attempt to unbind unitialize attr " + attr + " for store " + this.storeName);
-
-        this.__listeners[attr].splice(this.__listeners[attr].indexOf(handler));
-    },
-
-    hydrate: function hydrate(newState) {
-        for (var key in this.state) {
-            this.state[key] = newState[key];
-        }
-    },
-
-    dehydrate: function() {
-        return this.state;
-    },
-    
-    __notify: function notifyStateChangeHandler(name, value, oldValue) {
-        if (!this.__listeners[name])
-            return;
-
-        this.__listeners[name].forEach(function(handler) {
-            handler(null, value, oldValue);
-        }.bind(this));
-    }
-};
+var StoreStateBasePrototype = require("./store/proto");
 
 /**
  * Encapsulate state in a "read-only" object.
  */
-function createStoreState(props, observer) {
+function createStoreState(props, observer, strict) {
     var state = {};
 
     var keeper = Object.keys(props).reduce(function(stateKeeper, key) {
+        var def = props[key];
+        
         Object.defineProperty(stateKeeper, key, {
             enumerable: true,
             get: function() {
@@ -68,6 +18,15 @@ function createStoreState(props, observer) {
             },
 
             set: function(val) {
+                if(def.type) {
+                    var err = def.type(val);
+
+                    if( strict && err instanceof Error){
+                        console.error("Strict warning: setting %s with value %s failed due to typeConstraints", key, val);    
+                        return;
+                    }
+                }
+
                 var prevValue = state[key];
                 state[key] = val;
                 observer(key, val, prevValue);
@@ -129,7 +88,7 @@ function createStore(storeName, props) {
                 enumerable: true,
                 get: function() {
                     if (!state)
-                        state = createStoreState(props.stateProps, this.__notify.bind(this));
+                        state = createStoreState(props.stateProps, this.__notify.bind(this), props.strict);
 
                     return state;
                 }
